+++ mobile/android/docs/changelog.md
@@ -23,6 +23,15 @@ permalink: /changelog/
 * **feature-awesomebar**
  * Search engine suggestions will only be displayed if the user inputs at least 2 characters and matches the starting characters of the search engine name. [bug #1851012](https://bugzilla.mozilla.org/show_bug.cgi?id=1851012)
 
+**logins-storage**
+
+* Removed SQLCipher logins migration path for users from v95 and below
+
+* ⚠️ **This will cause a loss to existing logins if**:
+  * User is using the logins feature and has a version of v95 or below
+  * User is NOT syncing their logins with another device
+  * User plans to upgrade from v95 to v119 or above
+
 # 118.0
 * [Commits](https://github.com/mozilla-mobile/firefox-android/compare/releases_v117..releases_v118)
 * [Dependencies](https://github.com/mozilla-mobile/firefox-android/blob/releases_v118/android-components/plugins/dependencies/src/main/java/DependenciesPlugin.kt)